refactor: ChessRequest를 dto로 이동
